Added Like Search support to Query Parser

+++ rest-api/src/main/java/org/openams/rest/jpa/entity/TestScore.java
@@ -43,7 +43,7 @@ public class TestScore implements Serializable {
 
 	private String notes;
 
-	private int score;
+	private Float score;
 
 	@Temporal(TemporalType.TIMESTAMP)
 	@Column(name="start_dtt")
